refactor: extract element define macros
