+++ rusaint/src/webdynpro/element/action/link.rs
@@ -2,7 +2,7 @@ use std::{borrow::Cow, cell::OnceCell, collections::HashMap};
 
 use crate::webdynpro::{
     element::{
-        define_element_interactable,
+        macros::define_element_interactable,
         property::{HotkeyValue, Visibility},
         Interactable,
     },
