+++ rusaint/src/webdynpro/element/complex/sap_table/cell/hierarchical_cell.rs
@@ -10,7 +10,7 @@ use crate::webdynpro::element::{
         },
         SapTable,
     },
-    sub::define_subelement,
+    sub::macros::define_subelement,
     ElementDefWrapper,
 };
