+++ rusaint/src/webdynpro/element/complex/sap_table/cell/selection_cell.rs
@@ -3,8 +3,11 @@ use std::{borrow::Cow, cell::OnceCell};
 use scraper::Selector;
 
 use crate::webdynpro::element::{
-    complex::{sap_table::{property::SapTableCellType, SapTableDef}, SapTable},
-    sub::define_subelement,
+    complex::{
+        sap_table::{property::SapTableCellType, SapTableDef},
+        SapTable,
+    },
+    sub::macros::define_subelement,
     ElementDefWrapper,
 };
