+++ rusaint/src/webdynpro/element/layout/scrollbar.rs
@@ -1,6 +1,6 @@
 use std::{borrow::Cow, cell::OnceCell};
 
-use crate::webdynpro::element::{define_element_interactable, property::Visibility};
+use crate::webdynpro::element::{macros::define_element_interactable, property::Visibility};
 
 use self::property::ScrollDirection;
